Upgrade to webpack 2, fix hot reloading
