+++ web/.babelrc
@@ -1,5 +1,5 @@
 {
-    "presets": ["es2015", "stage-0", "react"],
+    "presets": [["es2015", { "modules": false }], "stage-0", "react"],
     "plugins": [
       "react-hot-loader/babel"
     ],
