+++ web/app/Main.js
@@ -3,8 +3,8 @@ if (!window.Intl) { // Safari polyfill
     	require.ensure(['intl'], require => {
     	window.Intl = require('intl');
         Intl.__addLocaleData(require("./assets/intl-data/en.json"));
-        require("App.jsx");
+        require("index.js");
     });
 } else {
-    require("App.jsx");
+    require("index.js");
 }
