+++ web/app/assets/index.html
@@ -41,7 +41,7 @@
     </main>
 
 <script src="app.js" charset="utf-8"></script>
-<script src="1.app.js" charset="utf-8"></script>
+<!-- <script src="1.app.js" charset="utf-8"></script> -->
 <script>
     var BTSW = window.BTSW || {};
     (function () {
