+++ web/app/assets/loader.js
@@ -1,7 +1,7 @@
 require("./stylesheets/app.scss");
-require("file?name=index.html!./index.html");
-require("file?name=favicon.ico!./favicon.ico");
-require("file?name=dictionary.json!common/dictionary_en.json");
+require("file-loader?name=index.html!./index.html");
+require("file-loader?name=favicon.ico!./favicon.ico");
+require("file-loader?name=dictionary.json-loader!common/dictionary_en.json");
 require("babel/polyfill");
 require("whatwg-fetch");
 require("indexeddbshim");
