+++ web/app/assets/locales/locales.js
@@ -1,7 +1,7 @@
-require("file?name=[name].[ext]!./locale-cn.json");
-require("file?name=[name].[ext]!./locale-de.json");
-require("file?name=[name].[ext]!./locale-en.json");
-require("file?name=[name].[ext]!./locale-es.json");
-require("file?name=[name].[ext]!./locale-fr.json");
-require("file?name=[name].[ext]!./locale-ko.json");
-require("file?name=[name].[ext]!./locale-tr.json");
+require("file-loader?name=[name].[ext]!./locale-cn.json");
+require("file-loader?name=[name].[ext]!./locale-de.json");
+require("file-loader?name=[name].[ext]!./locale-en.json");
+require("file-loader?name=[name].[ext]!./locale-es.json");
+require("file-loader?name=[name].[ext]!./locale-fr.json");
+require("file-loader?name=[name].[ext]!./locale-ko.json");
+require("file-loader?name=[name].[ext]!./locale-tr.json");
