+++ web/app/components/Exchange/exchange.scss
@@ -134,12 +134,6 @@ form.order-form {
     border-top: 1px solid #777;
     border-bottom: 1px solid #777;
   }
-  // .top-header {
-  //   border-top: 1px solid #777;
-  // }
-  // .bottom-header {
-  //   border-bottom: 1px solid #777;
-  // }
 }
 
 .final-row > td {
