+++ web/app/components/Wallet/BrainkeyInput.jsx
@@ -6,7 +6,7 @@ import Translate from "react-translate-component";
 var dictionary_set;
 
 if (__ELECTRON__) {
-    dictionary_set = new Set(require("json!common/dictionary_en.json").en.split(","));
+    dictionary_set = new Set(require("json-loader!common/dictionary_en.json").en.split(","));
 }
 
 export default class BrainkeyInput extends Component {
