+++ web/conf/webpack.config.js
@@ -2,22 +2,50 @@ var path = require("path");
 var webpack = require("webpack");
 var ExtractTextPlugin = require("extract-text-webpack-plugin");
 var Clean = require("clean-webpack-plugin");
-var git = require('git-rev-sync');
-require('es6-promise').polyfill();
+var git = require("git-rev-sync");
+require("es6-promise").polyfill();
 
 // BASE APP DIR
 var root_dir = path.resolve(__dirname, "..");
 
-// FUNCTION TO EXTRACT CSS FOR PRODUCTION
-function extractForProduction(loaders) {
-  return ExtractTextPlugin.extract("style", loaders.substr(loaders.indexOf("!")));
-}
-
 module.exports = function(options) {
     // console.log(options.prod ? "Using PRODUCTION options\n" : "Using DEV options\n");
     // STYLE LOADERS
-    var cssLoaders = "style-loader!css-loader!postcss-loader",
-      scssLoaders = "style!css!postcss-loader!sass?outputStyle=expanded";
+    var cssLoaders = [
+        {
+            loader: "style-loader"
+        },
+        {
+            loader: "css-loader"
+        },
+        {
+            loader: "postcss-loader",
+            options: {
+                plugins: []
+            }
+        }
+    ];
+
+    var scssLoaders =  [
+        {
+            loader: "style-loader"
+        },
+        {
+            loader: "css-loader"
+        },
+        {
+            loader: "sass-loader",
+            options: {
+                outputStyle: "expanded"
+            }
+        },
+        {
+            loader: "postcss-loader",
+            options: {
+                plugins: []
+            }
+        }
+    ];
 
     // DIRECTORY CLEANER
     var cleanDirectories = ["dist"];
@@ -27,7 +55,6 @@ module.exports = function(options) {
 
     // COMMON PLUGINS
     var plugins = [
-        new webpack.optimize.DedupePlugin(),
         new webpack.optimize.OccurrenceOrderPlugin(),
         new webpack.DefinePlugin({
             APP_VERSION: JSON.stringify(git.tag()),
@@ -37,12 +64,16 @@ module.exports = function(options) {
 
     if (options.prod) {
         // WRAP INTO CSS FILE
-        cssLoaders = extractForProduction(cssLoaders);
-        scssLoaders = extractForProduction(scssLoaders);
+        cssLoaders = [{loader: ExtractTextPlugin.extract(
+            {loader: ["style-loader", "css-loader", "postcss-loader"]}
+        )}];
+        scssLoaders = [{loader: ExtractTextPlugin.extract(
+            {loader: ["style-loader", "css-loader", "postcss-loader", "sass-loader?outputStyle=expanded"]}
+        )}];
 
         // PROD PLUGINS
         plugins.push(new Clean(cleanDirectories, {root: root_dir}));
-        plugins.push(new webpack.DefinePlugin({'process.env': {NODE_ENV: JSON.stringify('production')}}));
+        plugins.push(new webpack.DefinePlugin({"process.env": {NODE_ENV: JSON.stringify("production")}}));
         plugins.push(new ExtractTextPlugin("app.css"));
         if (!options.noUgly) {
             plugins.push(new webpack.optimize.UglifyJsPlugin({
@@ -65,7 +96,7 @@ module.exports = function(options) {
         // PROD OUTPUT PATH
         outputPath = path.join(root_dir, "dist");
     } else {
-        plugins.push(new webpack.optimize.OccurenceOrderPlugin());
+        // plugins.push(new webpack.optimize.OccurenceOrderPlugin());
         plugins.push(new webpack.DefinePlugin({"process.env": {NODE_ENV: JSON.stringify("development")}}));
         plugins.push(new webpack.HotModuleReplacementPlugin());
         plugins.push(new webpack.NoErrorsPlugin());
@@ -89,22 +120,28 @@ module.exports = function(options) {
             sourceMapFilename: "[name].js.map"
         },
         devtool: options.prod ? "cheap-module-source-map" : "eval",
-        debug: options.prod ? false : true,
         module: {
-            loaders: [
+            rules: [
                 {
                     test: /\.jsx$/,
                     include: [path.join(root_dir, "app"), path.join(root_dir, "node_modules/react-foundation-apps"), "/home/sigve/Dev/graphene/react-foundation-apps"],
-                    loaders: options.prod ? ["babel-loader"] : ["babel-loader?cacheDirectory"]
+                    use: [
+                        {
+                            loader: "babel-loader",
+                            options: {
+                                cacheDirectory: options.prod ? false : true
+                            }
+                        }
+                    ]
                 },
                 {
                     test: /\.js$/,
                     exclude: [/node_modules/],
                     loader: "babel-loader",
-                    query: {compact: false, cacheDirectory: true}
+                    options: {compact: false, cacheDirectory: true}
                 },
                 {
-                    test: /\.json/, loader: "json",
+                    test: /\.json/, loader: "json-loader",
                     exclude: [
                         path.resolve(root_dir, "lib/common"),
                         path.resolve(root_dir, "app/assets/locales")
@@ -112,38 +149,57 @@ module.exports = function(options) {
                 },
                 { test: /\.coffee$/, loader: "coffee-loader" },
                 { test: /\.(coffee\.md|litcoffee)$/, loader: "coffee-loader?literate" },
-                { test: /\.css$/, loader: cssLoaders },
+                {
+                    test: /\.css$/,
+                    use: cssLoaders
+                },
+
+                // var cssLoaders = "style-loader!css-loader!postcss-loader",
+                //   scssLoaders =  "style-loader!css-loader!postcss-loader!sass-loader?outputStyle=expanded";
+
                 {
                     test: /\.scss$/,
-                    loader: scssLoaders
+                    use: scssLoaders
                 },
                 { test: /\.png$/, loader: "url-loader?limit=100000", exclude:[
                     path.resolve(root_dir, "app/assets/asset-symbols")
                 ] },
                 { test: /\.woff$/, loader: "url-loader?limit=100000&mimetype=application/font-woff" },
                 { test: /.*\.svg$/, loaders: ["svg-inline-loader", "svgo-loader"] },
-                { test: /\.md/, loader: 'html?removeAttributeQuotes=false!remarkable' }
-            ],
-            postcss: function () {
-                return [precss, autoprefixer];
-            }
+                {
+                    test: /\.md/,
+                    use: [
+                        {
+                            loader: "html-loader",
+                            options: {
+                                removeAttributeQuotes: false
+                            }
+                        },
+                        {
+                            loader: "remarkable-loader",
+                            options: {
+                                preset: "full",
+                                typographer: true
+                            }
+                        }
+                    ]
+                }
+            ]
         },
         resolve: {
-            root: [path.resolve(root_dir, "./app"), path.resolve(root_dir, "./lib")],
-            extensions: ["", ".js", ".jsx", ".coffee", ".json"],
-            modulesDirectories: ["node_modules"],
-            fallback: [path.resolve(root_dir, "./node_modules")]
+            modules: [
+                path.resolve(root_dir, "./app"),
+                path.resolve(root_dir, "./lib"),
+                "node_modules"
+            ],
+            extensions: [".js", ".jsx", ".coffee", ".json"],
+            // fallback: [path.resolve(root_dir, "./node_modules")]
         },
         resolveLoader: {
-            root: path.join(root_dir, "node_modules"),
-            fallback: [path.resolve(root_dir, "./node_modules")]
+            modules: [path.join(root_dir, "node_modules")],
+            // fallback: [path.resolve(root_dir, "./node_modules")]
         },
-        plugins: plugins,
-        root: outputPath,
-        remarkable: {
-            preset: "full",
-            typographer: true
-        }
+        plugins: plugins
     };
 
     // if(options.prod) config.entry.vendors = [
