+++ web/lib/common/localStorage.js
@@ -1,9 +1,9 @@
 // Localstorage
-import ls, {ls_key_exists} from './localStorageImpl';
+import ls, {ls_key_exists} from "./localStorageImpl";
 
 if (null===ls) throw "localStorage is required but isn't available on this platform";
 
-module.exports = (key) => {
+const localStorage = (key) => {
 
     var STORAGE_KEY = key;
 
@@ -32,4 +32,6 @@ module.exports = (key) => {
             return ls_key_exists(STORAGE_KEY + key, ls);
         }
     };
-}
+};
+
+export default localStorage;
